+++ pareg_core/src/from_read.rs
@@ -1,4 +1,10 @@
-use crate::{reader::Reader, ArgError, Result};
+use std::{
+    ffi::OsString,
+    net::{Ipv4Addr, SocketAddrV4},
+    path::PathBuf,
+};
+
+use crate::{parsef_part, reader::Reader, ArgError, ParseFArg, Result};
 
 /// Trait similar to [`crate::FromArg`]. Difference is that this may parse only
 /// part of the input.
@@ -140,4 +146,91 @@ macro_rules! impl_from_arg_str_with_read {
         {
         }
     };
+}
+
+impl FromRead for bool {
+    fn from_read(r: &mut Reader) -> Result<(Self, Option<ArgError>)> {
+        let (c, _) = char::from_read(r)?;
+        match c {
+            't' => {
+                r.expect("rue")?;
+                Ok((true, None))
+            }
+            'f' => {
+                r.expect("alse")?;
+                Ok((false, None))
+            }
+            c => Err(r.err_parse(format!(
+                "Expected `true` or `false`, but there is `{c}`"
+            ))),
+        }
+    }
+}
+
+impl FromRead for char {
+    fn from_read(r: &mut Reader) -> Result<(Self, Option<ArgError>)> {
+        let Some(c) = r.next().transpose()? else {
+            return Err(r.err_parse("Expected character."));
+        };
+        Ok((c, None))
+    }
+}
+
+impl FromRead for String {
+    fn from_read(r: &mut Reader) -> Result<(Self, Option<ArgError>)> {
+        let mut res = String::new();
+        r.read_all(&mut res)?;
+        Ok((res, None))
+    }
+}
+
+impl FromRead for PathBuf {
+    fn from_read(r: &mut Reader) -> Result<(Self, Option<ArgError>)> {
+        let mut res = String::new();
+        r.read_all(&mut res)?;
+        Ok((res.into(), None))
+    }
+}
+
+impl FromRead for OsString {
+    fn from_read(r: &mut Reader) -> Result<(Self, Option<ArgError>)> {
+        let mut res = String::new();
+        r.read_all(&mut res)?;
+        Ok((res.into(), None))
+    }
+}
+
+impl FromRead for Ipv4Addr {
+    fn from_read(r: &mut Reader) -> Result<(Self, Option<ArgError>)> {
+        let mut c: (u8, u8, u8, u8) = Default::default();
+        let r = parsef_part(
+            r,
+            [
+                ParseFArg::Arg(&mut c.0),
+                ParseFArg::Str(".".into()),
+                ParseFArg::Arg(&mut c.1),
+                ParseFArg::Str(".".into()),
+                ParseFArg::Arg(&mut c.2),
+                ParseFArg::Str(".".into()),
+                ParseFArg::Arg(&mut c.3),
+            ],
+        )?;
+        Ok((Ipv4Addr::new(c.0, c.1, c.2, c.3), r))
+    }
+}
+
+impl FromRead for SocketAddrV4 {
+    fn from_read(r: &mut Reader) -> Result<(Self, Option<ArgError>)> {
+        let mut adr: Ipv4Addr = Ipv4Addr::LOCALHOST;
+        let mut port: u16 = 0;
+        let r = parsef_part(
+            r,
+            [
+                ParseFArg::Arg(&mut adr),
+                ParseFArg::Str(":".into()),
+                ParseFArg::Arg(&mut port),
+            ],
+        )?;
+        Ok((SocketAddrV4::new(adr, port), r))
+    }
 }